refactor(backend): DB 관련 파일을 database/ 폴더로 이동
prismaClient.ts와 dbPath.ts를 루트 및 utils/에서 database/ 폴더로 이동하여
DB 인프라 파일을 한 곳에 모았다.
